feat: add list component

+++ src/modules/nova/primitives/f32/f32.component.h
@@ -1,14 +1,15 @@
-#include "i32.imports.h"
+#include "f32.imports.h"
 #include <stdlib.h>
 annotation(import, modules.nova, nv)
 
+
 annotation(type)
 struct T {
-  int32_t value;
+  float value;
 };
 
 annotation(constructor)
-static T create(int32_t value){
+static T create(float value){
   T self = nv.checkAllocation(malloc(sizeof(*self)));
   self->value = value;
   return self;
@@ -20,11 +21,11 @@ static void destroy(T self){
 }
 
 annotation(method)
-static int32_t get(T self){
+static float get(T self){
   return self->value;
 }
 
 annotation(method)
-static void set(T self, int32_t value){
+static void set(T self, float value){
   self->value = value;
 }
