+++ src/modules/nova/primitives/i64/i64.component.h
@@ -1,14 +1,14 @@
-#include "i32.imports.h"
+#include "i64.imports.h"
 #include <stdlib.h>
 annotation(import, modules.nova, nv)
 
 annotation(type)
 struct T {
-  int32_t value;
+  int64_t value;
 };
 
 annotation(constructor)
-static T create(int32_t value){
+static T create(int64_t value){
   T self = nv.checkAllocation(malloc(sizeof(*self)));
   self->value = value;
   return self;
@@ -20,11 +20,11 @@ static void destroy(T self){
 }
 
 annotation(method)
-static int32_t get(T self){
+static int64_t get(T self){
   return self->value;
 }
 
 annotation(method)
-static void set(T self, int32_t value){
+static void set(T self, int64_t value){
   self->value = value;
 }
